// nxtstep-interview-engine/src/types/interview.types.ts
export type InterviewState =
  | 'INIT'
  | 'PREP'
  | 'GENERATE_Q'
  | 'AWAIT_ANSWER'
  | 'PROCESS_ANSWER'
  | 'EVALUATE'
  | 'DECIDE_FOLLOWUP'
  | 'GENERATE_FU'
  | 'AWAIT_FU_ANSWER'
  | 'LOOP'
  | 'TERMINATE'
  | 'AGGREGATE'
  | 'COMPLETE';